requirements.txt
whisper
sounddevice
numpy
scipy
gTTS
openai
requests
tqdm
transformers
torch
huggingface-hub
python-dotenv